requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2
imbalanced-learn>=0.11
scikit-optimize>=0.9
shap>=0.42
umap-learn>=0.5
joblib>=1.3
tqdm>=4.65
upsetplot>=0.9
adjustText>=0.8
missforest>=2.0
openpyxl>=3.1
pyyaml>=6.0
statsmodels>=0.14
